requirements.txt
absl-py==0.5.0
appnope==0.1.0
astor==0.7.1
astroid==2.0.4
backcall==0.1.0
beautifulsoup4==4.6.3
bleach==3.0.2
certifi==2018.8.24
chardet==3.0.4
clarifai==2.3.2
configparser==3.5.0
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
entrypoints==0.2.3
future==0.16.0
gast==0.2.0
googleapis-common-protos==1.5.3
grpcio==1.15.0
h5py==2.8.0
idna==2.7
ipykernel==5.1.0
ipython==7.0.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==3.0.0
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.0
numpy==1.15.2
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prometheus-client==0.4.1
prompt-toolkit==2.0.6
protobuf==3.6.1
ptyprocess==0.6.0
Pygments==2.2.0
pylint==2.1.1
pyparsing==2.2.2
python-dateutil==2.7.3
python-igraph==0.7.1.post6
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.2
requests==2.19.1
scikit-learn==0.20.0
scipy==1.1.0
selenium==3.14.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
tensorboard==1.11.0
tensorflow==1.11.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
tqdm==4.26.0
traitlets==4.3.2
typed-ast==1.1.0
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.10.11
